Renamed StudyInfo folder to TutkinnonOsa
