+++ shared/components/TukinnonOsa/CompetenceAquirementTitle.tsx
