+++ shared/components/TukinnonOsa/Competences.tsx
