+++ shared/components/TukinnonOsa/Details.tsx
