+++ shared/components/TukinnonOsa/LearningEvent.tsx
