+++ shared/components/TukinnonOsa/MobileCompetences.tsx
