+++ shared/components/TukinnonOsa/Objectives.tsx
