+++ shared/components/TukinnonOsa/Period.tsx
