+++ shared/components/TukinnonOsa/RequirementsAndDeviations.tsx
