+++ shared/components/TukinnonOsa/Shared.tsx
