+++ shared/components/TukinnonOsa/StudyInfoHelpers.ts
